Added operators for flechter types.
