+++ integrations/vivado_hls/src/fletcher/components/meta.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "types.h"
+
+/// @brief RecordBatch metadata.
+struct RecordBatchMeta
+{
+    f_base_length_type length;
+};
