+++ integrations/vivado_hls/src/fletcher/components/operators/arith.h
@@ -1,0 +1,50 @@
+#pragma once
+
+#include "../packet.h"
+#include "../nullable.h"
+
+#define OP +
+
+#include "base_arith.inc"
+#include "base_unary.inc"
+
+#define OP -
+
+#include "base_arith.inc"
+#include "base_unary.inc"
+
+#define OP *
+
+#include "base_arith.inc"
+
+#define OP /
+
+#include "base_arith.inc"
+
+#define OP %
+
+#include "base_arith.inc"
+
+#define OP &
+
+#include "base_arith.inc"
+
+#define OP |
+
+#include "base_arith.inc"
+
+#define OP ^
+
+#include "base_arith.inc"
+
+#define OP ~
+
+#include "base_unary.inc"
+
+#define OP <<
+
+#include "base_arith.inc"
+
+#define OP >>
+
+#include "base_arith.inc"
