+++ integrations/vivado_hls/src/fletcher/components/operators/base_logical_derived.inc
@@ -1,0 +1,17 @@
+template <typename T>
+bool operator!=(TYPEL lhs, TYPER rhs) {
+    return !(rhs == lhs);
+}
+
+template <typename T>
+bool operator>(TYPEL lhs, TYPER rhs) {
+    return rhs < lhs;
+}
+template <typename T>
+bool operator<=(TYPEL lhs, TYPER rhs) {
+    return !(rhs < lhs);
+}
+template <typename T>
+bool operator>=(TYPEL lhs, TYPER rhs) {
+    return !(lhs < rhs);
+}
