+++ integrations/vivado_hls/src/fletcher/components/operators/indecrement.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "../packet.h"
+#include "../nullable.h"
+
+#define OP ++
+
+#include "base_instance.inc"
+
+#define OP --
+
+#include "base_instance.inc"
