+++ integrations/vivado_hls/src/fletcher/components/packet_base.h
@@ -1,0 +1,14 @@
+#pragma once
+
+/// @brief Base packet containing count, dvalid and last.
+struct f_packet_base
+{
+    bool dvalid = true;
+    bool last = false;
+
+    f_packet_base() = default;
+
+    virtual ~f_packet_base() = default;
+
+    f_packet_base(bool _dvalid, bool _last) : dvalid(_dvalid), last(_last) {}
+};
